requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.3
django-dotenv==1.4.2
django-filter==23.2
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gprof2dot==2022.7.29
pycodestyle==2.10.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
